[#156] ref: WriteViewController TableView 리팩토링 - DataSource 별도 분리 - 셀 구성 메서드 분리
