+++ ThingLog/View/TableVeiwCell/WriteTextFieldCell.swift
@@ -25,23 +25,6 @@ final class WriteTextFieldCell: UITableViewCell {
         return textField
     }()
 
-    var placeholder: String? {
-        didSet {
-            textField.attributedPlaceholder = NSAttributedString(string: placeholder ?? "",
-                                                                 attributes: [
-                                                                    NSAttributedString.Key.foregroundColor: SwiftGenColors.gray3.color,
-                                                                    NSAttributedString.Key.font: UIFont.Pretendard.body1
-                                                                 ])
-        }
-    }
-
-    var keyboardType: UIKeyboardType = .default {
-        didSet {
-            textField.keyboardType = keyboardType
-            textField.isSelection = keyboardType == .numberPad ? false : true
-        }
-    }
-
     var isEditingSubject: PublishSubject<Bool> = PublishSubject<Bool>()
     var textValueSubject: PublishSubject<String?> = PublishSubject<String?>()
 
@@ -65,6 +48,17 @@ final class WriteTextFieldCell: UITableViewCell {
 
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+
+    /// 텍스트 필드의 키보드 타입과 플레이스 홀더를 구성한다.
+    func configure(keyboardType: UIKeyboardType, placeholder: String) {
+        textField.keyboardType = keyboardType
+        textField.isSelection = keyboardType == .numberPad ? false : true
+        textField.attributedPlaceholder = NSAttributedString(string: placeholder ?? "",
+                                                             attributes: [
+                                                                NSAttributedString.Key.foregroundColor: SwiftGenColors.gray3.color,
+                                                                NSAttributedString.Key.font: UIFont.Pretendard.body1
+                                                             ])
     }
 }
